vcard_support.c: use charset_decode_mimebody() rather than sasl_decode64()

+++ imap/vcard_support.h
@@ -59,8 +59,8 @@ extern struct buf *vcard_as_buf(struct vparse_card *vcard);
 extern struct vparse_card *record_to_vcard(struct mailbox *mailbox,
                                            const struct index_record *record);
 
-extern unsigned vcard_prop_decode_value(struct vparse_entry *prop,
-                                        struct buf *value,
-                                        struct message_guid *guid);
+extern size_t vcard_prop_decode_value(struct vparse_entry *prop,
+                                      struct buf *value,
+                                      struct message_guid *guid);
 
 #endif /* VCARD_SUPPORT_H */
